Move list_datasets test to proper file
